refactor : Refactor stream code convention
- 스트림 API를 사용할 때 가독성을 위해서 한줄에 점 하나씩만 찍도록 수정하였습니다.
- 이외에도 일관성이 없는 부분들을 수정하였습니다.

+++ src/main/java/org/prgrms/kdtspringorder/io/enums/implementation/CommandType.java
@@ -21,7 +21,7 @@ public enum CommandType {
   }
 
   public boolean hasOption(String option) {
-    int cnt = (int) Arrays.stream(this.options)
+    long cnt = Arrays.stream(this.options)
         .filter(possibleOption -> possibleOption.getOptionName().equals(option))
         .count();
     return cnt > 0;
